fix: Insert Flow Change To Stake Set, Manager Docker Flow

+++ src/main/java/com/sotatek/cardano/job/schedules/CardanoCliSchedule.java
@@ -7,7 +7,6 @@ import com.sotatek.cardano.job.event.KafkaRegistryEvent;
 import com.sotatek.cardano.job.service.interfaces.CardanoCliService;
 import java.math.BigInteger;
 import java.util.Objects;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -21,18 +20,16 @@ import org.springframework.util.ObjectUtils;
 public class CardanoCliSchedule {
 
   public static final String BYRON = "byron";
-  public static final String SHELLEY = "Shelley";
   private final QueryCli queryCli;
   private final CardanoCliService cardanoCliService;
   private final ApplicationEventPublisher publisher;
   private String epochStakeTopicId;
-  private AtomicBoolean afterByronFirstTime = new AtomicBoolean(Boolean.TRUE);
-  private AtomicInteger failEpoch = new AtomicInteger(0);
+  public static final AtomicInteger  failEpoch = new AtomicInteger(0);
 
   public CardanoCliSchedule(QueryCli queryCli,
       CardanoCliService cardanoCliService,
       ApplicationEventPublisher publisher,
-      @Value("${kafka.topics.epochStake.name}") String epochStakeTopicId) {
+      @Value("${kafka.topics.epoch.name}") String epochStakeTopicId) {
     this.queryCli = queryCli;
     this.cardanoCliService = cardanoCliService;
     this.publisher = publisher;
@@ -40,7 +37,7 @@ public class CardanoCliSchedule {
   }
 
 
-  @Scheduled(fixedDelayString = "${jobs.cardano-cli.delay}")
+  @Scheduled(initialDelay = 300, fixedDelayString = "${jobs.cardano-cli.delay}")
   public void getCliEpoch() {
     var streamEpochResult = cardanoCliService.runExec(queryCli.getTip());
     if (Objects.isNull(streamEpochResult)) {
@@ -58,15 +55,8 @@ public class CardanoCliSchedule {
 
       NodeInfo nodeInfo = cardanoCliService.getNodeInfo();
 
-      if (afterByronFirstTime.get()
-          && tip.getEra().equals(SHELLEY)) {
-        afterByronFirstTime.set(Boolean.FALSE);
-        nodeInfo.setEpochNo(tip.getEpoch());
-        log.info("current epoch {}, era {}", tip.getEpoch(), tip.getEra());
-        return;
-      }
-
-      if (tip.getEpoch() - nodeInfo.getEpochNo() > BigInteger.ONE.longValue()) {
+      if (nodeInfo.getEpochNo() != BigInteger.ZERO.intValue() &&
+          tip.getEpoch() - nodeInfo.getEpochNo() > BigInteger.ONE.longValue()) {
         cardanoCliService.removeContainer();
         return;
       }
@@ -79,6 +69,7 @@ public class CardanoCliSchedule {
 
           if (failEpoch.get() != BigInteger.ZERO.intValue() && failEpoch.get() <= tip.getEpoch()) {
             publisher.publishEvent(new KafkaRegistryEvent(epochStakeTopicId, Boolean.TRUE));
+            failEpoch.set(BigInteger.ZERO.intValue());
           }
         }
         nodeInfo.setEpochNo(tip.getEpoch());
